Tab reorganization
Did a reorganization of sections of the site based on advice of a
student who knows nothing of the event. He wants sponsors first so he
can see what companies are going to be at the event. Also wants to know
what speakers are going. Will redo color scheme later.